chore: add clockMode field
