+++ apps/meteor/client/views/account/accessibility/hooks/useAcessibilityPreferencesValues.ts
@@ -6,6 +6,7 @@ export type AccessibilityPreferencesData = {
 	themeAppearence?: ThemePreference;
 	fontSize?: FontSize;
 	fontSizePreference?: FontSize;
+	clockMode?: 0 | 1 | 2;
 	hideUsernames?: boolean;
 	hideRoles?: boolean;
 };
@@ -13,12 +14,14 @@ export type AccessibilityPreferencesData = {
 export const useAccessiblityPreferencesValues = (): AccessibilityPreferencesData => {
 	const themeAppearence = useUserPreference<ThemePreference>('themeAppearence') || 'auto';
 	const fontSize = useUserPreference<FontSize>('fontSize') || '100%';
+	const clockMode = useUserPreference<0 | 1 | 2>('clockMode') ?? 0;
 	const hideUsernames = useUserPreference<boolean>('hideUsernames');
 	const hideRoles = useUserPreference<boolean>('hideRoles');
 
 	return {
 		themeAppearence,
 		fontSize,
+		clockMode,
 		hideUsernames,
 		hideRoles,
 	};
